% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_renderFlow.R
\name{.renderSingle}
\alias{.renderSingle}
\title{.renderSingle}
\usage{
.renderSingle(
  rmdFile,
  sampleId,
  data,
  sampleInfo,
  reference,
  interpretation,
  saveDir,
  report_format,
  product_code
)
}
\description{
.renderSingle
}
