% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_resultFigure.R
\name{.trans2showvalueL}
\alias{.trans2showvalueL}
\title{.trans2showvalueL}
\usage{
.trans2showvalueL(value, lower_limit)
}
\arguments{
\item{value}{numeric, detect value}

\item{lower_limit}{numeric}
}
\value{
trans_value numeric
}
\description{
convert detect value; only lower_limit
}
