% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_resultFigure.R
\name{.trans2showvalueLMU}
\alias{.trans2showvalueLMU}
\title{.trans2showvalueLMU}
\usage{
.trans2showvalueLMU(value, lower_limit, middle_limit, upper_limit)
}
\arguments{
\item{value}{numeric, detect value}

\item{lower_limit}{numeric}

\item{middle_limit}{numeric}

\item{upper_limit}{numeric}
}
\description{
convert detect value; both lower_limit, middle_limit, upper_limit
}
