% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_resultFigure.R
\name{.trans2showvalueLU}
\alias{.trans2showvalueLU}
\title{.trans2showvalueLU}
\usage{
.trans2showvalueLU(value, lower_limit, upper_limit, minimal, maximal)
}
\arguments{
\item{value}{numeric, detect value}

\item{lower_limit}{numeric}

\item{upper_limit}{numeric}
}
\value{
trans_value numeric
}
\description{
convert detect value; both lower_limit and upper_limit
}
