% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_resultFigure.R
\name{.trans2showvalueU}
\alias{.trans2showvalueU}
\title{.trans2showvalueU}
\usage{
.trans2showvalueU(value, upper_limit)
}
\arguments{
\item{value}{numeric, detect value}

\item{upper_limit}{numeric}
}
\value{
trans_value numeric
}
\description{
convert detect value; only upper_limit
}
