% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_nutriMeta.R
\name{interpret_nutriMeta}
\alias{interpret_nutriMeta}
\title{interpret_nutriMeta}
\usage{
interpret_nutriMeta(
  data,
  reference,
  interpret_text,
  product_label,
  ch_compound_levels,
  collapse = "<br />&emsp;&emsp;"
)
}
\arguments{
\item{data}{detect data}

\item{reference}{limits info; tibble}

\item{interpret_text}{interpret text; tibble}

\item{product_label}{product type; character}

\item{ch_compound_levels}{list}

\item{collapse}{collapse; character}
}
\description{
interpret_nutriMeta
}
