% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cardio.R
\name{interprete_DX2057}
\alias{interprete_DX2057}
\title{interprete_DX2057}
\usage{
interprete_DX2057(data, reference)
}
\arguments{
\item{data}{tibble}

\item{reference}{tibble}
}
\value{
text
}
\description{
interprete_DX2057
}
