% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cardio.R
\name{interprete_DX2059}
\alias{interprete_DX2059}
\title{interprete_DX2059}
\usage{
interprete_DX2059(data, reference)
}
\arguments{
\item{data}{tibble}

\item{reference}{tibble}
}
\value{
list
}
\description{
interprete_DX2059
}
