% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_resultFigure.R
\name{trans2show_cardio_ggplot}
\alias{trans2show_cardio_ggplot}
\title{trans2show_cardio_ggplot}
\usage{
trans2show_cardio_ggplot(
  data,
  reference,
  ch_compound_levels,
  barWidth,
  textSize
)
}
\arguments{
\item{data}{dataframe, detect value}

\item{ch_compound_levels}{ch_compound_levels}

\item{barWidth}{barWidth}

\item{textSize}{textSize}

\item{limitsInfo}{dataframe, limits info}
}
\value{
ggplot object
}
\description{
generate showing figure for HCY
}
