% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_resultFigure.R
\name{trans2show_ceram_ggplot}
\alias{trans2show_ceram_ggplot}
\title{trans2show_ceram_ggplot}
\usage{
trans2show_ceram_ggplot(score)
}
\arguments{
\item{data}{dataframe, detect value}

\item{HCY_limitsInfo}{dataframe, limits info}
}
\value{
ggplot object
}
\description{
generate showing figure for ceram
}
