% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_resultFigure.R
\name{trans2show_nutrition_ggplot}
\alias{trans2show_nutrition_ggplot}
\title{trans2show_nutrition_ggplot}
\usage{
trans2show_nutrition_ggplot(
  data,
  reference,
  barWidth,
  textSize,
  ch_compound_levels
)
}
\arguments{
\item{data}{dataframe, detect value}

\item{reference}{dataframe, limits info}

\item{barWidth}{barWidth}

\item{textSize}{textSize}

\item{ch_compound_levels}{list chr}
}
\value{
ggplot object
}
\description{
generate showing figure for HCY
}
